requirements.txt
streamlit
torch
transformers
Pillow
nltk
gTTS